updated amount from double to BigDecimal

+++ src/main/java/com/revoltcode/account/cmd/command/DepositFundsCommand.java
@@ -1,9 +1,12 @@
 package com.revoltcode.account.cmd.command;
 
 import com.revoltcode.cqrs.core.command.BaseCommand;
+import lombok.AllArgsConstructor;
 import lombok.Data;
+
+import java.math.BigDecimal;
 
 @Data
 public class DepositFundsCommand extends BaseCommand {
-    private double amount;
+    private BigDecimal amount;
 }
